Fine, I'll do it myself
okay, so I dug into this problem and it seems that the ai is really bad at setting up environments, so i did it myself. I figure that setting up the environment isn't really that big a deal once you know what you're doing, so it shouldn't be a setback.